requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic-settings==2.6.1
python-telegram-bot==21.6
SQLAlchemy==2.0.48
typing-extensions>=4.14.1
python-dotenv==1.0.1

